added benchmark options

+++ src/A2G.BitPermutationLab/Benchmarking/BenchmarkSelectionOptions.cs
@@ -3,5 +3,5 @@ namespace A2G.BitPermutationLab.Benchmarking;
 public sealed record BenchmarkSelectionOptions(
     WeightingProfileKind WeightingProfile,
     int? ScenarioBudget,
-    int SamplingSeed,
+    ulong SamplingSeed,
     bool IncludeRequiredBaselines = true);
